Move response handling to utils file
