+++ exec/utils.go
@@ -3,10 +3,14 @@ package exec
 import (
 	"context"
 	"fmt"
+	"io"
 	"os"
 	"os/exec"
 	"runtime"
+	"strings"
+	"sync"
 
+	"github.com/mitchellh/go-linereader"
 	"github.com/turbot/steampipe-plugin-sdk/v5/plugin"
 )
 
@@ -85,4 +89,60 @@ func prepareCommand(ctx context.Context, d *plugin.QueryData, _ *plugin.HydrateD
 
 	return cmd, nil
 
+}
+
+func copyUIOutput2(ctx context.Context, d *plugin.QueryData, r io.Reader, doneCh chan<- struct{}) error {
+	plugin.Logger(ctx).Warn("listRemoteCommandResult", "ctx_done", "copyUIOutput2 starting...")
+	defer close(doneCh)
+	lr := linereader.New(r)
+	i := 1
+	for line := range lr.Ch {
+		d.StreamListItem(ctx, outputRow{Line: line, LineNumber: i, Stream: "stdout"})
+		i = i + 1
+		//o.Output(line)
+	}
+	plugin.Logger(ctx).Warn("listRemoteCommandResult", "ctx_done", "copyUIOutput2 done")
+	return nil
+}
+
+func copyUIOutput3(ctx context.Context, d *plugin.QueryData, r io.Reader, isLineByLine bool, isError bool) error {
+	plugin.Logger(ctx).Warn("listRemoteCommandResult", "ctx_done", "copyUIOutput3 starting...")
+
+	stream := "stdout"
+	if isError {
+		stream = "stderr"
+	}
+
+	if isLineByLine {
+		lr := linereader.New(r)
+		i := 1
+		for line := range lr.Ch {
+			d.StreamListItem(ctx, outputRow{Line: line, LineNumber: i, Stream: stream})
+			i = i + 1
+		}
+	} else {
+		buf := new(strings.Builder)
+		n, _ := io.Copy(buf, r)
+		if n == 0 {
+			return nil
+		}
+		d.StreamListItem(ctx, outputRow{Line: buf.String(), LineNumber: 1, Stream: stream})
+	}
+
+	plugin.Logger(ctx).Warn("listRemoteCommandResult", "ctx_done", "copyUIOutput3 done")
+	return nil
+}
+
+func copyUIOutput4(ctx context.Context, d *plugin.QueryData, r io.Reader, wg *sync.WaitGroup) error {
+	plugin.Logger(ctx).Warn("listRemoteCommandResult", "ctx_done", "copyUIOutput2 starting...")
+	defer wg.Done()
+	lr := linereader.New(r)
+	i := 1
+	for line := range lr.Ch {
+		d.StreamListItem(ctx, outputRow{Line: line, LineNumber: i, Stream: "stdout"})
+		i = i + 1
+		//o.Output(line)
+	}
+	plugin.Logger(ctx).Warn("listRemoteCommandResult", "ctx_done", "copyUIOutput2 done")
+	return nil
 }
